Placeholder Jump Behavior First Pass
This patch implements placholder navigation. It implements a state
machine to track the users behavior and trigger jumping at the right
time ( in theory ).

It improves the users typing flow and feels more like Xcode, which is
good for navigating SDK APIS.

Design:
- In completion mode, typing a key other than <Tab> or <Down> visible
navigates to the next placeholder.
- The Tab key navigates between placeholders.

This is incrediblly hard to get right. Perhaps in the future  there
could be an ability to turn off the jumping, and leave it with the
original keystroke based stuff if this causes too many problems.

It manually checks if the user is inside of a string: It's sort of
hacky, and syntax isn't working here with Apple's syntax files.

Good enough for me.

Additionally Update ICMD

- Changes for placeholders
- Stability issues on the server

Tested on
Vim version 8.0.567