Coexist: iCompleteMe client namespace
This patch allows iCompleteMe and YouCompleteMe to operate in the same
process. I need this because I'll be using YouCompleteMe for some C++,
C, and ObjC coding in the same projects as swift.

There is a single vim python process, for all plugins, so it isn't
possible to have ICM and YCM adding the same paths, or there will be
collisions.

YCM shares code with YCMD by adding the paths of YCM to the global
import path.

- Apply import renames
- Move around folders

Additionally, namespace global variables to prevent collisions.

+ Update ICMD for missing file