Disable/Enable Ycm If needed
It looks like Ycm autocmds are causing issues with Icms.

Disable/Enable the movement behavior when needed.

This should be preferred to disablng Ycm altogehter. When working in a
project with multiple languages, it is desirable to having completion
working for all languages.

If this is still causing issues, consider disabling YCM altogether, i.e.
by conditionally setting g:loaded_youcompleteme = 1